requirements.txt
pydantic-settings==2.8.1
langchain==0.3.21
langchain-openai==0.3.1
langgraph==0.3.21
langsmith==0.3.19
psycopg-binary==3.2.6
python-dotenv==1.1.0
langchain-community==0.3.20
tavily-python==0.5.4
fastapi==0.115.12
psycopg2-binary==2.9.10
passlib==1.7.4
python-jose==3.4.0
python-multipart==0.0.20
bcrypt==4.3.0
langchain-tavily==0.1.5
langgraph-checkpoint-postgres==2.0.19
wikipedia==1.4.0
arxiv==2.2.0
langchain-chroma==0.2.3